redesigned package hash function
There are two hashes now. one for folder contents and one for the current state of the extension in memory. This way if any sys.path addresses are added to the extension due to addition of library extensions, asmkmaker will failed the hash check and will recompile an assembly for the extension with new search paths