Revised output stream flush data
up to this point, the output stream would write output as they're handed to the Write method. This meant that any output bigger than 1024 bytes would be separated into chunks and the Write method would wrap them in a <div> but since the data was not complete, the html output would not look right. With this revision, the output stream works like a real stream and collected the data being written to the output and creates the html code only when flushed. This ensures that a single large stream of data is wrapped correctly in <div>