Using a different approach to overcome output size limit
Assuming the buffer size will be a factor of 1024 (1K), the output stream won't print the output if the incoming buffer is completely full. It assumes that when buffer is full, the input stream must be larger than the buffer. 
This obviously means that any output that is exactly a factor of 1024 will be printed with the next print statement.