Converted all explicit str() to unicode() in markdown module
For compatibility with IronPython unicode
markdown module check python version and creates a wrapper around the unicode element that is designed for python 2.
IronPython however works differently. unicode is a wrapper for the standard str in IronPython and:
custom_unicode = unicode
is really:
custom_unicode = str
which makes this code throw an exception:
custom_unicode(u'中國哲學書電子化計劃')
but the unicode keyword itself does not raise any exceptions:
unicode(u'中國哲學書電子化計劃')     # this is fine