COREUPDATE: Added lightscopes option to the pyRevitLoader
Loader uses a lot of memory launching pyRevit and creating all the extensions and commands. It also needs to have the lightscopes option similar to the pyRevit individual commands

Multiple tests were performed to see if a single engine can be reused between the commands but results were unsuccessful. Each script is a like a module run inside the engine and thus all the loaded modules will be shared between scripts and this breaks many of the scripts or modules that rely on self configuration on load time based on open documents, etc. e.g. rpw module

Also once the io streams are set for an engine, it seems that reassigning them to a new window has no effect.