Added support for directory state cache files
Extension developers can create a hash file storing the extension calculated cache value. This helps with increasing performance over extensions shared over network drives.
If hash file is provided for an extension, pyRevit will use that hash instead of calculating it.
NOTE:
This means the developers need to update cache when adding, removing, adjusting the layout, ... that changes the state of their ui (other than icons since they're not included in hash calcs). Algorithmic changes to scripts should not change the cache value since the link to file name is still the same.