Decided to move on to a better output html rendere
The WebBrowser control uses the installed IE activex control on the machine. Apparently there are memory leaks involved with the event handlers on this control. After a long battle decided to move on to a more advanced html renderer and probably a WPF window. That'll give more flexibility and less headache for furture upgardes