Revised standard library system
pylib was the IronPython standard library
pyRevit engine is now using the ZIP file of the library inside the PyRevitLoader module. This allows the executor to use whatever IronPython engine that the PyRevitLoader is compiled against as the executing engine. The lib folder has also been renamed to site-packages to better follow the python stdlib rules.