New tool to set active workset from selection's workset
Second try at creating a new tool I built in dynamo first and ported to pyRevit.
I started from the create workset by link script to adapt it in a breeze.
I hope that is better than the previous tool

Not sure if there is a prefered location for the tool