Modification to Import/Export Family Configs (#1117)
* fixed host adder and added 2020.2.3

* added 2020.1.1 to hosts

* Modified import and export buttons to handle shared parameters.
The original behavior would assume all parameters were not shared.
This commit detects shared parameters, and if they exist, creates
another file and uses it to import them into Revit.

* Resolves the issue of the Load Classification type being treated as a string.
The yaml file assumes the Load Classification Type to be a string, so when using
the original import button, the Load Classification in Revit would not populate.
The export script attaches a string to the beginning of the Load Classification
Type in the yaml file. When the import script comes across the notifier string,
it imports the value as a Load Classification Type. A more general solution is
necessary for other types of parameters that use a ElementId.

Co-authored-by: Ehsan Iran-Nejad <eirannejad@gmail.com>