Sort revisions before printing (#1583)
The previous script outputs the revision list straight from a FilteredElementCollector, so the list of revisions are not necessarily in numerical order. This change gathers the revision numbers and revision data in a rev_data list prior to adding them to the rev_table string, that way they can be sorted first.