Add .NET 8 Runtime Check to Installer Scripts
Implemented a check for the .NET 8 Runtime in the installation scripts for pyRevit, pyRevit CLI, and pyRevit Admin. If the required runtime is not found, users are prompted to download it from the official Microsoft site. This ensures that the application has the necessary dependencies before proceeding with the installation.
Fixes [Bug]: Installer does not check for .NET "core" runtime #2730