Add CodeDependencies inclusion in installer scripts
Updated the pyRevit, pyRevit CLI, and pyRevit Admin installer scripts to include the "CodeDependencies.iss" file. This change ensures that the necessary dependencies for .NET 8 are properly managed during the installation process, enhancing the overall robustness and reliability of the installation. Additionally, the architectures installation mode has been reinstated to x64 for consistency across scripts.