Update __init__.py
Rewrite of the print_table output method.

This is in response to the need for greater control over column widths and other formatting options.

The method is rewritten to directly output an HTML table instead of the original markdown approach. It introduces more CSS options through classes. This in turn allows different formatting options to be specified for each table in the same output window (which was limited in the original).

The new funcionality can be optionally called with new kwargs. If these are not specified, the original functionality of the old print_table is maintained. 

It adds some data validation to check that arguments of type list are of the same length.

Finally, this contribution should male it easier to later implement HTML colspan options to make more flexible tables.